refactor: remove unused logger and validation utilities
- Deleted `logger.ts` and `validation.ts` as they were not utilized in the codebase.
- Replaced debug statements with `console.log` for better visibility during execution.
- Updated imports in `commits.ts`, `release.ts`, `workspace.ts` to reflect these changes.
