+++ src/workspace.ts
@@ -3,14 +3,13 @@ import type {
   FindWorkspacePackagesOptions,
   PackageJson,
   PackageUpdateOrder,
+  VersionUpdate,
   WorkspacePackage,
 } from "./types";
 import { readFile } from "node:fs/promises";
 import { join } from "node:path";
-import { createDebugger } from "./logger";
 import { run } from "./utils";
-
-const debug = createDebugger("ucdjs:release-scripts:workspace");
+import { createVersionUpdate } from "./version";
 
 interface RawProject {
   name: string;
@@ -71,7 +70,7 @@ export async function findWorkspacePackages(
     const packageJson: PackageJson = JSON.parse(content);
 
     if (!shouldIncludePackage(packageJson, options)) {
-      debug?.(`Excluding package ${rawProject.name}`);
+      console.log(`Excluding package ${rawProject.name}`);
       continue;
     }
 
@@ -215,4 +214,54 @@ export function getAllDependents(
 
   visit(packageName);
   return result;
+}
+
+/**
+ * Pure function: Determine which packages need updates due to dependency changes
+ *
+ * When a package is updated, all packages that depend on it should also be updated.
+ * This function calculates which additional packages need patch bumps.
+ *
+ * @param updateOrder - Packages in topological order with their dependency levels
+ * @param directUpdates - Packages with direct code changes
+ * @returns All updates including dependent packages
+ */
+export function createDependentUpdates(
+  updateOrder: Array<{ package: WorkspacePackage; level: number }>,
+  directUpdates: VersionUpdate[],
+): VersionUpdate[] {
+  const allUpdates = [...directUpdates];
+  const updatedPackages = new Set(directUpdates.map((u) => u.package.name));
+
+  // Process packages in dependency order
+  for (const { package: pkg } of updateOrder) {
+    // Skip if already updated
+    if (updatedPackages.has(pkg.name)) {
+      continue;
+    }
+
+    // Check if any workspace dependencies are being updated
+    if (hasUpdatedDependencies(pkg, updatedPackages)) {
+      // This package needs a patch bump because its dependencies changed
+      allUpdates.push(createVersionUpdate(pkg, "patch", false));
+      updatedPackages.add(pkg.name);
+    }
+  }
+
+  return allUpdates;
+}
+
+/**
+ * Pure function: Check if a package has any updated dependencies
+ */
+export function hasUpdatedDependencies(
+  pkg: WorkspacePackage,
+  updatedPackages: Set<string>,
+): boolean {
+  const allDeps = [
+    ...pkg.workspaceDependencies,
+    ...pkg.workspaceDevDependencies,
+  ];
+
+  return allDeps.some((dep) => updatedPackages.has(dep));
 }
